Revert "Revert #cb4815e (#516)"
This reverts commit b18f1e23c58027578534ddb93ac2058388a6e3b2.
