+++ next/src/components/sections/BranchesSection.tsx
@@ -1,11 +1,9 @@
-import { useQuery } from '@tanstack/react-query'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
 
 import BranchCard from '@/src/components/common/Card/BranchCard'
 import SectionContainer from '@/src/components/layout/Section/SectionContainer'
 import SectionHeader from '@/src/components/layout/Section/SectionHeader'
-import { client } from '@/src/services/graphql'
 import { BranchesSectionFragment } from '@/src/services/graphql/api'
 import { isDefined } from '@/src/utils/isDefined'
 import { useGetFullPath } from '@/src/utils/useGetFullPath'
@@ -22,19 +20,11 @@ const BranchesSection = ({ section }: Props) => {
   const { i18n } = useTranslation()
   const locale = i18n.language
 
-  const { title, text, branches, showAll } = section
+  const { title, text, branches } = section
 
   const { getFullPath } = useGetFullPath()
 
-  // TODO consider optimalizing so that we don't fetch this much when showAll is false
-  const { data: allBranches } = useQuery({
-    queryFn: () => client.Branches({ locale }),
-    queryKey: ['branches', locale],
-  })
-
-  const branchesToRender =
-    // eslint-disable-next-line unicorn/no-array-callback-reference
-    (showAll ? allBranches?.branches : branches)?.data.filter(isDefined) ?? []
+  const filteredBranches = branches?.data.filter(isDefined) ?? []
 
   return (
     // TODO padding-y should probably be managed by the SectionContainer
@@ -42,7 +32,7 @@ const BranchesSection = ({ section }: Props) => {
       <div className="flex flex-col items-start gap-6 lg:gap-12">
         <SectionHeader title={title} text={text} />
         <ul className="flex flex-col gap-4 self-stretch lg:grid lg:grid-cols-3 lg:items-start lg:gap-8">
-          {branchesToRender
+          {filteredBranches
             .map((branch, index) => {
               if (!branch.attributes) return null
 
@@ -51,8 +41,8 @@ const BranchesSection = ({ section }: Props) => {
                 <li key={index} className="h-full [&>*]:h-full">
                   <BranchCard
                     title={branch.attributes.title}
-                    address={branch.attributes.address ?? ''}
-                    linkHref={getFullPath(branch) ?? '#'}
+                    address={branch.attributes.address}
+                    linkHref={getFullPath(branch)}
                   />
                 </li>
               )
