+++ next/src/services/graphql/queries/_header-sections.gql
@@ -50,6 +50,14 @@ fragment PickupDayHeaderSection on ComponentHeaderSectionsPickupDay {
   }
 }
 
+fragment BranchMapHeaderSection on ComponentHeaderSectionsBranchMap {
+  branches {
+    data {
+      ...BranchEntity
+    }
+  }
+}
+
 fragment HeaderSections on PageHeaderDynamicZone {
   __typename
 
@@ -69,15 +77,11 @@ fragment HeaderSections on PageHeaderDynamicZone {
     ...GalleryHeaderSection
   }
 
-  #  ... on ComponentHeaderSectionsIcon {
-  #    ...IconHeaderSection
-  #  }
-
   ... on ComponentHeaderSectionsPickupDay {
     ...PickupDayHeaderSection
   }
 
-  ... on ComponentSectionsBranchesMap {
-    ...BranchesMapSection
+  ... on ComponentHeaderSectionsBranchMap {
+    ...BranchMapHeaderSection
   }
 }
