+++ next/src/services/graphql/queries/_page-sections.gql
@@ -68,17 +68,6 @@ fragment ImageAndTextOverlappedSection on ComponentSectionsImageAndTextOverlappe
 fragment BranchesSection on ComponentSectionsBranches {
   title
   text
-  showAll
-  branches {
-    data {
-      ...BranchEntity
-    }
-  }
-}
-
-fragment BranchesMapSection on ComponentSectionsBranchesMap {
-  title
-  text
   branches {
     data {
       ...BranchEntity
@@ -389,10 +378,6 @@ fragment PageSections on PageSectionsDynamicZone {
 
   ... on ComponentSectionsBranches {
     ...BranchesSection
-  }
-
-  ... on ComponentSectionsBranchesMap {
-    ...BranchesMapSection
   }
 
   ... on ComponentSectionsWorkshops {
