+++ strapi/src/api/page/content-types/page/schema.json
@@ -66,7 +66,7 @@
         "header-sections.image",
         "header-sections.side-image",
         "header-sections.pickup-day",
-        "sections.branches-map"
+        "header-sections.branch-map"
       ],
       "max": 1,
       "pluginOptions": {
@@ -103,8 +103,7 @@
         "sections.opening-times",
         "sections.board-members",
         "sections.vacancies",
-        "sections.global-search",
-        "sections.branches-map"
+        "sections.global-search"
       ],
       "pluginOptions": {
         "i18n": {
